Tweak strings.ParseStringIntoWords

+++ src/strings/ParseStringIntoWords.java
@@ -29,7 +29,7 @@ public final class ParseStringIntoWords {
         if (s.isEmpty()) {
             return words;
         }
-        for (int i = 0; i < (int)s.length(); ++i) {
+        for (int i = 0; i < s.length(); ++i) {
             String firstWord = s.substring(0, i + 1);
             if (!dictionary.contains(firstWord)) {
                 continue;
@@ -38,8 +38,8 @@ public final class ParseStringIntoWords {
             List<String> remainderParse =
                     parseStringIntoWords(remainder, dictionary);
             if (remainderParse == null) {
-                // Maybe trying a longer first word
-                // will allow us to parse the remainder.
+                // We failed to parse the remainder.
+                // Let's try a longer first word.
                 continue;
             }
             words.add(firstWord);
